Add a button for convenient exporting.
The Mu Collider toolbar tab has been renamed to Mu Tools, and an export
panel added. The export panel has a button that exports the currently
selected root object (same as the file menu item) but sets the default name
of the file to the name of the root object.