Get armature-based hierarchies working for imports.
If "Force Armature" is set when importing, or the .mu has skinned mesh
renderers, the hierarchy is created using an armature with each transform
(except root, that's the armature) as a bone. Non-deformable objects are
children of the relevant bones. Skinned meshes will be attached
appropriately when I figure out how.

Lights (and probably cameras) not yet oriented correctly.