Add a tool to clear parent inverse cleanly.
That is, when the parent inverse matrix is cleared, the object keeps its
local (and thus world) transforms; basis transform (UI visible) is updated
to suit. This is necessary because .mu files expect pure local transforms,
but the location, rotation and scale properties of a blender object use the
basis matrix, as do animations in blender. Thus, to get correct behavior in
KSP, the basis and local matrices need to match, with an identity parent
inverse matrix.