Add my mass export script.
This exports all root objects that have children and are enabled for
rendering. The names of the .mu files are taken from the root objects (with
any .nnn stripped), and all images referenced by mu shaders are exported to
png. .cfg.in templates are parsed too, but that's part of the core export
process.

blender -noaudio --background project.blend -P mass-export.py