Use dual armatures to handle bind poses
Unity's (and really, probably any engine's) SkinnedMeshRenderer's
bindPoses contains only those bones that contribute to the deformation,
so much of the hierarchy may be missing and is impossible to reconstruct
(infinite solutions). Instead, use two armatures to deform the mesh: one
that is set up using the bind pose data (and thus has only those bones)
and a second that acts as the main armature, with all the bones inferred
from the game object hierarchy. The poses of the bind pose armature
bones are set from the rest poses of the corresponding bones in the main
armature, thus giving the mesh its initial deformation as specified by
the game object hierarchy. The bind pose armature is made a child of the
main armature in order to avoid any potential issues, and both armatures
are added as modifiers to the mesh, with the bind pose armature added
first so it does the initial deformation while the main armature does
the animation.