Create a new mesh object when exporting
It turns out the result of to_mesh() is horribly ephemeral and a new(?)
function bpy.data.meshes.new_from_object() should be used when
obtaining modifier-applied mesh data. It might be slower (unknown, not
measured) but that's not very relevant compared to "it works". DG meshes
are no longer corrupted.