Add a flag to force colliders onto separate GOs
While having a collider on the same game object as the vis mesh is more
efficient, sometimes having it separate is required. With the "Separate"
flag, single colliders can optionally be forced onto a separate game
object. The flag is automatically set if the collider is the only
component on the object when importing.