Avoid creating dup curves for some redundant data
The problem comes from animated properties on materials that are shared
across multiple objects: the materials are unique to the model, but the
animation data is per object. With this fix, the material gets redundant
NLA strips which will need to be cleaned up (better than throwing an
exception while importing, though).

Fixes #23