Handle incorrect edge splits
Because edge-touch detection has an epsilon (currently excessive, but
it's likely to come up with any non-zero epsilon), there are times when
the point is outside the triangle on two faces instead of just one (or
none as the point may actually be on the inside of the triangle, but
that is usually not a problem). This results in one of the two
replacement triangles being flipped. Detecting the flip and then
replicating the non-split code generates correct triangles.