Implement basics of fitted capsule colliders
Right now, the size is such that the capsule fits *inside* the larger
dimensions of the equivalent rectangular prism based on the moment of
inertia of the point cloud of the selected verts, but it *is* oriented
correctly with the minimum possible rotation (ie, the principle axis is
correctly selected). Next step will be to fit the capsule so it fully
encloses all the points but with the minimum size.