Lock down the environment passed to eval
While working on Diamond Grid I found that I could do a lot more with
the expressions passed to eval than I expected. This got me to wondering
about eval's safety... which does not exist by default. However, it
turns out to be not so difficult to lock down: just block dangerous
builtins (or really, allow ones that seem safe) to take care of 99% of
the holes leaving only crazy stuff with class accessors, but that's
handled by rejecting any string with double underscores (and the fact
that eval, exec and attribute functions (getattr etc) are blocked takes
care of string manipulation shenanigans to get around simple checks for
double underscores).