[collider] Find the eigenvectors much more robustly
I suspect Cayley-Hamilton works only for non-repeated (actually, I think
I read that somewhere, but finding info on why I used it proved
troublesome). It's still used for the 3-value case. However, for for the
one value case, the identity matrix is returned since all basis sets are
equivalent, and for two value cases, the nullspaces of the two matrices
are merged.

Also, matrix rows are no longer swapped: that proved to be incorrect,
though ensuring the rows point along their respective +ve axes does seem
to be correct.

There may still be some issues with the actual capsule calculation, but
it is now correctly oriented for random orientations, and the code
doesn't fail when the point cloud is rotationally symmetric.