[cfgnode] Rework to use a class for values
Now, values and subnodes are stored directly in the parent node's lists
rather than in name,value,line tuples (name and line are kept in the
value or node object). The idea is nodes and values will be storable in
a single list (not done yet), thus preserving their order when
manipulating them (eg, for templates). It will also allow a new object
for storing block comments (also not implemented). However, what is
implemented is nodes and values will write any single-line comments
attached to them.