[export] Correct prop hierarchy assumption
While KSP makes props direct children of the Internal, there should be
no need for them to be such in blender. This does mean there may be
useless empties in the exported model, but if that proves to be a
problem, something can be done then.

The prop's frame is made relative to the Internal root frame, so they
should be placed correctly regardless of intervening transformations.