[export] Add support for MODEL{} nodes
This allows the same sub-model to be reused in multiple parts,
internals, and props. This is done by marking the "root" object of the
sub-model as a model (in its model type field). When those objects are
encountered while exporting, instead of continuing down the tree, a
suitable MODEL{} node is injected into the output cfg file.

Collection instance handling has been fixed up such that if there's only
one root object in the collection, the collection itself is NOT
exported, but rather the root object replaces the collection in the
exported model. This is done before the above model type handling, so if
that root object is a model, a MODEL node will be emitted, otherwise the
collection's hierarchy is embedded in the .mu.

While there may be some offset issue, the above should go a long way to
getting prop collections to work nicely.