[import,export] Use ∧ to separate mu and model names
That is, the ∧ and all characters after it in the name are stripped on
export. ∧ takes priority over the .nnn check, but the .nnn check is
still done because manually adding ∧ to every object name in a large
file doesn't seem reasonable. However, imported objects have ∧ appened
to the name in the .mu file before creating the object so blender's .nnn
suffix will come after the ∧ and thus get stripped, but any such
suffixes (eg, .875 for part sizes) will be unaffected.

Fixes #22